+++ frontend/storybook/src/components/entry/admin/images/meta/lists/list-hover.ts
@@ -1,7 +1,7 @@
 import "@elements/core/lists/list-vertical";
 import "@elements/core/inputs/checkbox";
 export default {
-  title: 'Admin / Images / Lists',
+  title: 'Entry / Admin / Images / Meta / Lists',
 }
 
 export const ListHover = ({label}) => {
