+++ frontend/storybook/src/components/entry/admin/images/meta/lists/school.ts
@@ -6,7 +6,7 @@ import {SCHOOL_OPTIONS} from "~/mock/meta";
 
 
 export default {
-  title: 'Admin / Images / Lists',
+  title: 'Entry / Admin / Images / Meta / Lists',
 }
 
 interface Props {
