+++ frontend/storybook/src/components/entry/admin/images/meta/sections/image-label-right.ts
@@ -1,12 +1,12 @@
-import {ImageStyle} from "~/components/entry/admin/images/lists/image-style";
-import {Age} from "~/components/entry/admin/images/lists/age";
-import {Stream} from "~/components/entry/admin/images/lists/stream";
+import {ImageStyle} from "~/components/entry/admin/images/meta/lists/image-style";
+import {Age} from "~/components/entry/admin/images/meta/lists/age";
+import {Stream} from "~/components/entry/admin/images/meta/lists/stream";
 import "@elements/core/inputs/text-underline";
 import "@elements/core/lists/list-two-column";
 import "@elements/core/images/ji";
 
 export default {
-  title: 'Admin/Images/Settings',
+  title: 'Entry/Admin/Images/Sections',
 }
 
 const STR_USED = "To be used only for";
